Creation of CartTable component, renders all CartItems
